csharp-advanced-topics/UsingLazyInCSharp/HowToUseLazyInCSharpTests/LazyInitializationUnitTest.cs
using HowtoUseLazyInCSharp;
using Microsoft.Data.SqlClient;
using Microsoft.VisualStudio.TestTools.UnitTesting;

namespace HowToUseLazyInCSharpTests
{
    [TestClass]
    public class LazyInitializationUnitTest
    {
        [TestMethod]
        public void GivenLazyObjectExample_WhenRun_ThenLazyObjectIsCreated()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act
            var result = initialization.RunLazyObjectExample();

            // Assert
            Assert.IsNotNull(result);
        }

        [TestMethod]
        public void GivenLazyExpensiveObjectExample_WhenRun_ThenLazyExpensiveObjectIsCreated()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act
            var result = initialization.RunLazyExpensiveObjectExample();

            // Assert
            Assert.IsNotNull(result);
            Assert.IsInstanceOfType(result, typeof(ExpensiveObject));
        }

        [TestMethod]
        public void GivenLazyObjectWithValueExample_WhenRun_ThenLazyObjectWithValueIsCreated()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act
            var result = initialization.RunLazyObjectWithValueExample();

            // Assert
            Assert.IsNotNull(result);
            Assert.IsInstanceOfType(result, typeof(ExpensiveObject));
        }

        [TestMethod]
        public void GivenLazyObjectWithExceptionHandlingExample_WhenRun_ThenExceptionIsHandled()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act & Assert
            Assert.ThrowsException<Exception>(() => initialization.RunLazyObjectWithExceptionHandlingExample());
        }

        [TestMethod]
        public void GivenInitialOverheadExample_WhenRun_ThenObjectIsInitialized()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act
            initialization.RunInitialOverheadExample();

            // Assert: No exception is thrown
        }

        [TestMethod]
        public void GivenMemoryConsumptionExample_WhenRun_ThenObjectIsInitialized()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act
            initialization.RunMemoryConsumptionExample();

            // Assert: No exception is thrown
        }

        [TestMethod]
        public void GivenThreadSynchronizationOverheadExample_WhenRun_ThenObjectIsInitialized()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act
            initialization.RunThreadSynchronizationOverheadExample();

            // Assert: No exception is thrown
        }

        [TestMethod]
        public void GivenLazyInitializationExample_WhenRun_ThenExpensiveObjectIsInitialized()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act
            ExpensiveObject expensiveObject = initialization.RunLazyInitializationExample();

            // Assert
            Assert.IsNotNull(expensiveObject);
            // Add more assertions based on the behavior of the ExpensiveObject class
        }

        [TestMethod]
        public void GivenLazyConfigurationExample_WhenRun_ThenConfigurationIsInitialized()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act
            Configuration config = initialization.RunLazyConfigurationExample();

            // Assert
            Assert.IsNotNull(config);
            
        }

        [TestMethod]
        public void GivenEagerInitializationExample_WhenRun_ThenLoggerIsInitialized()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act & Assert
            initialization.RunEagerInitializationExample();
            
        }

        [TestMethod]
        public void GivenLazyCacheExample_WhenRun_ThenCacheIsInitialized()
        {
            // Arrange
            var initialization = new LazyInitialization();

            // Act
            Cache cache = initialization.RunLazyCacheExample();

            // Assert
            Assert.IsNotNull(cache);
            
        }
    }
}
